algorand/requirements.txt
pyteal==0.27.0
beaker-pyteal==1.1.1
python-dotenv==1.0.1
algosdk==2.7.0
